# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Timetable API",
        "description": "API documentation for the Timetable application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/timetable-change-requests": {
            "get": {
                "tags": [
                    "Timetable Change Requests"
                ],
                "summary": "Get list of timetable change requests (Admin only)",
                "description": "Returns list of timetable change requests",
                "operationId": "getChangeRequestsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TimetableChangeRequest"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Timetable Change Requests"
                ],
                "summary": "Store new timetable change request",
                "description": "Returns change request data",
                "operationId": "storeChangeRequest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "generated_timetable_id",
                                    "message"
                                ],
                                "properties": {
                                    "generated_timetable_id": {
                                        "type": "integer"
                                    },
                                    "message": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TimetableChangeRequest"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/timetable-change-requests/{id}": {
            "get": {
                "tags": [
                    "Timetable Change Requests"
                ],
                "summary": "Get change request information (Admin or owner)",
                "description": "Returns change request data",
                "operationId": "getChangeRequestById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Change Request id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TimetableChangeRequest"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Timetable Change Requests"
                ],
                "summary": "Update existing change request (Admin only)",
                "description": "Returns updated change request data",
                "operationId": "updateChangeRequest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Change Request id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "approved",
                                            "rejected"
                                        ]
                                    },
                                    "admin_response": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TimetableChangeRequest"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Timetable Change Requests"
                ],
                "summary": "Delete existing change request (Admin only)",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteChangeRequest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Change Request id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GeneratedTimetable": {
                "title": "Generated Timetable",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "user_id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "timetable": {
                        "type": "object"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
            "TimetableChangeRequest": {
                "title": "Timetable Change Request",
                "required": [
                    "generated_timetable_id",
                    "message"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "user_id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "generated_timetable_id": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "default": "pending",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected"
                        ]
                    },
                    "admin_response": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "generated_timetable": {
                        "$ref": "#/components/schemas/GeneratedTimetable"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "is_admin": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            }
        }
    },
...
}